refactor mousePosition
